fix: make role change notification an independent component
