+++ packages/roomkit-react/src/Prebuilt/components/Notifications/PeerNotifications.tsx
@@ -1,10 +1,7 @@
 import { useEffect } from 'react';
 import { HMSNotificationTypes, useHMSNotifications } from '@100mslive/react-sdk';
-import { useUpdateRoomLayout } from '../../provider/roomLayoutProvider';
 // @ts-ignore: No implicit Any
 import { ToastBatcher } from '../Toast/ToastBatcher';
-// @ts-ignore: No implicit Any
-import { ToastManager } from '../Toast/ToastManager';
 // @ts-ignore: No implicit Any
 import { useSetSubscribedChatSelector, useSubscribedNotifications } from '../AppData/useUISettings';
 // @ts-ignore: No implicit Any
@@ -14,7 +11,6 @@ const notificationTypes = [
   HMSNotificationTypes.PEER_JOINED,
   HMSNotificationTypes.PEER_LEFT,
   HMSNotificationTypes.NAME_UPDATED,
-  HMSNotificationTypes.ROLE_UPDATED,
 ];
 
 export const PeerNotifications = () => {
@@ -22,7 +18,6 @@ export const PeerNotifications = () => {
   const isPeerJoinSubscribed = useSubscribedNotifications(SUBSCRIBED_NOTIFICATIONS.PEER_JOINED);
   const isPeerLeftSubscribed = useSubscribedNotifications(SUBSCRIBED_NOTIFICATIONS.PEER_LEFT);
   const [selectedPeer, setPeerSelector] = useSetSubscribedChatSelector(CHAT_SELECTOR.PEER);
-  const updateRoomLayoutForRole = useUpdateRoomLayout();
 
   useEffect(() => {
     if (!notification?.data) {
@@ -47,15 +42,6 @@ export const PeerNotifications = () => {
       case HMSNotificationTypes.NAME_UPDATED:
         console.log(notification.data.id + ' changed their name to ' + notification.data.name);
         return;
-      case HMSNotificationTypes.ROLE_UPDATED: {
-        if (notification.data?.isLocal && notification.data?.roleName) {
-          ToastManager.addToast({
-            title: `You are now a ${notification.data.roleName}`,
-          });
-          updateRoomLayoutForRole?.(notification.data.roleName);
-        }
-        return;
-      }
       default:
         return;
     }
